extract reloadDataAsFullObjects to shared utility
